+++ lib/songle-song-uri-parser.rb
@@ -1,17 +1,19 @@
 # -*- coding: utf-8 -*-
 require "songle/song_uri"
-require "songle/song_uri/mp3_song_uri"
-require "songle/song_uri/numeric_song_uri"
+
+require "songle/song_uri/invalid_song_uri_error"
+
 require "songle/song_uri/nn_song_uri"
 require "songle/song_uri/nn_scheme_song_uri"
 require "songle/song_uri/nn_short_song_uri"
+
 require "songle/song_uri/sc_song_uri"
 require "songle/song_uri/sc_scheme_song_uri"
+
 require "songle/song_uri/yt_song_uri"
 require "songle/song_uri/yt_scheme_song_uri"
 require "songle/song_uri/yt_short_song_uri"
-require "songle/song_uri/invalid_song_uri_error"
 
-module ::SongleSongURIParser
-  VERSION = "1.1.1"
-end
+require "songle/song_uri/numeric_song_uri"
+
+require "songle/song_uri/mp3_song_uri"