+++ lib/songle/song_uri/numeric_song_uri.rb
@@ -16,12 +16,12 @@ module ::Songle::SongURI
     ##
     # @constructor
     #
-    def initialize query, options
-      super(query, options)
+    def initialize query_string, options
+      super(options)
 
       @source_host = nil
       @source_path = nil
-      @source_id   = $3 if query =~ URI_REGEX
+      @source_id   = $3 if query_string =~ URI_REGEX
     end
 
     ##