+++ lib/songle/song_uri/sc_song_uri.rb
@@ -26,15 +26,15 @@ module ::Songle::SongURI
     ##
     # @constructor
     #
-    def initialize query, options
-      super(query, options)
+    def initialize query_string, options
+      super(options)
 
       @source_host = SOURCE_HOST
-      @source_path = $1 if query =~ URI_REGEX
-      @source_id   = $2 if query =~ URI_REGEX
+      @source_path = $1 if query_string =~ URI_REGEX
+      @source_id   = $2 if query_string =~ URI_REGEX
 
       if @source_id.nil? || @source_id.strip.empty?
-        raise ::Songle::SongURI::InvalidSongURIError.new(query)
+        raise ::Songle::SongURI::InvalidSongURIError.new(query_string)
       end
     end